+++ settings.gradle.kts
@@ -56,7 +56,8 @@ if (scalaVersion == "2.12") {
   project(":spark-connector:spark-3.3").projectDir = file("spark-connector/v3.3/spark")
   project(":spark-connector:spark-runtime-3.3").projectDir = file("spark-connector/v3.3/spark-runtime")
   // flink only support scala 2.12
-  include("flink-connector")
+  include("flink-connector:flink")
+  include("flink-connector:flink-runtime")
 }
 include("spark-connector:spark-3.4", "spark-connector:spark-runtime-3.4", "spark-connector:spark-3.5", "spark-connector:spark-runtime-3.5")
 project(":spark-connector:spark-3.4").projectDir = file("spark-connector/v3.4/spark")
@@ -66,3 +67,7 @@ project(":spark-connector:spark-runtime-3.5").projectDir = file("spark-connector
 include("web")
 include("docs")
 include("integration-test-common")
+include("flink-connector:flink")
+findProject(":flink-connector:flink")?.name = "flink"
+include("flink-connector:flink-runtime")
+findProject(":flink-connector:flink-runtime")?.name = "flink-runtime"
